--- app/javascript/react/src/components/modal.css
@@ -1,62 +1,0 @@
-.modal {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 80vw; 
-  max-width: 800px;
-  height: 80vh; 
-  background: rgba(255, 255, 255, 1.00);
-  border-radius: 15px;
-}
-
-.modal-overlay {
-    position: fixed;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.5);
-}
-
-.modal::before {
-  content: "";
-  position: fixed;
-  top: 60px;
-  left: 4%;
-  width: 54vw;
-  height: 65vh;
-  background: rgba(236, 234, 234, 1.00);
-  border-radius: 15px;
-  z-index: 1;
-}
-
-.modal-text {
-  position: fixed;
-  top: 90px;
-  left: 17%;
-  transform: translate(-50%, -50%);
-  color: #000; 
-  font-size: 14px;
-  text-align: center;
-  z-index: 2;
-}
-
-.modal button {
-  background: none;
-  border: none;
-  cursor: pointer;
-  margin-top: 3%;
-  margin-left: 3%;
-  margin-right: 2%;
-}
-
-.modal button img {
-  position: absolute;
-  width: 25px;
-  height: 17px;
-}
-
-.modal button.close-button {
-  position: absolute;
-  right: 25px;
-}
